requirements.txt
transformers
Pillow
torch
torchvision
torchaudio
requests
gradio